fix for sandboxed install

+++ mkTextDomains.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-  for F in "${1}"/*.mo ; do
-    D=${F%.mo}/LC_MESSAGES/ ; mkdir -p "${D}" ; mv -f "${F}" "${D}/${2}.mo"
+  for F in "${INSTALL_ROOT}/${1}/"*.mo ; do
+    if [ -f "${F}" ] ; then
+      D=${F%.mo}/LC_MESSAGES/ ; mkdir -p "${D}" ; mv -f "${F}" "${D}/${2}.mo"
+    fi
   done
 
 #eof
